# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Directory API",
        "description": "API documentation for the Cart API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Retrieve a list of books",
                "description": "Retrieve a paginated list of books with filters and sorting.",
                "operationId": "e3ca1243c40f73fae90719ef44209035",
                "parameters": [
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "description": "Search by title, description, authors.last_name, authors.first_name or authors.patronymic.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[author]",
                        "in": "query",
                        "description": "Filter by authors slug.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by book fields.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "-created_at",
                                "title",
                                "-title",
                                "publication_date",
                                "-publication_date"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Include related authors.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "authors"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book list retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Create a book",
                "description": "Create a new book.",
                "operationId": "cea0e0738c55976e73681eba8cb71b26",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookFullResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/books/{book}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Retrieve book details",
                "description": "Retrieve detailed information about a specific book.",
                "operationId": "0ef12c3dc7339459157697e8efb88c4e",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "the-tragical-history-of-hamlet-prince-of-denmark"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookFullResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Book not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Remove a book",
                "description": "Remove the specified book.",
                "operationId": "9eda054341169792f0183f7fbe5ff89c",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "the-tragical-history-of-hamlet-prince-of-denmark"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted successfully."
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Book not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Book"
                ],
                "summary": "Update a book",
                "description": "Update the specified book.",
                "operationId": "3c4f25fa03441abe8a539d459cf8dfa1",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "the-tragical-history-of-hamlet-prince-of-denmark"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookFullResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Book not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthorResponse": {
                "title": "Author Response",
                "description": "Data of a specific author.",
                "required": [
                    "id",
                    "last_name",
                    "first_name",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the author.",
                        "type": "integer",
                        "example": 1
                    },
                    "last_name": {
                        "description": "Lastname of the author.",
                        "type": "string",
                        "example": "Shevchenko"
                    },
                    "first_name": {
                        "description": "Firstname of the author.",
                        "type": "string",
                        "example": "Taras"
                    },
                    "patronymic": {
                        "description": "Patronymic of the author.",
                        "type": "string",
                        "example": "Grigorievich",
                        "nullable": true
                    },
                    "slug": {
                        "description": "Slug of the author.",
                        "type": "string",
                        "example": "shevchenko-taras-grigorievich"
                    },
                    "birth_date": {
                        "description": "Birth date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1814-03-09",
                        "nullable": true
                    },
                    "death_date": {
                        "description": "Death date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1861-03-10",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Biography of the author.",
                        "type": "string",
                        "example": "Ukrainian poet, artist, and thinker. Born into a family of serfs, he gained his freedom through a ransom. His work became the foundation of modern Ukrainian literature.",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Image url of the author.",
                        "type": "string",
                        "format": "uri",
                        "example": "http://loclhost:8080/storage/authors/author1.png",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CreateBookRequest": {
                "title": "Create Book Request",
                "description": "Request payload for creation a new book.",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the book.",
                        "type": "string",
                        "example": "The Tragical History of Hamlet, Prince of Denmark",
                        "maxLength": 255
                    },
                    "slug": {
                        "description": "Slug of the book.",
                        "type": "string",
                        "pattern": "^[a-z0-9-]+$",
                        "example": "the-tragical-history-of-hamlet-prince-of-denmark",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "description": {
                        "description": "Description of the book.",
                        "type": "string",
                        "example": "The story follows Prince Hamlet of Denmark, who returns home to find his father (the King) dead and his uncle, Claudius, newly crowned and married to Hamlet's mother, Gertrude. After the Ghost of his father appears and reveals that Claudius murdered him, Hamlet is consumed by a quest for revenge.",
                        "nullable": true,
                        "maxLength": 5000
                    },
                    "publication_date": {
                        "description": "Publication date of the book.",
                        "type": "string",
                        "format": "date",
                        "example": "1603-01-01",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateBookRequest": {
                "title": "Update Book Request",
                "description": "Request payload for updating a new book.",
                "required": [],
                "properties": {
                    "title": {
                        "description": "Title of the book.",
                        "type": "string",
                        "example": "The Tragical History of Hamlet, Prince of Denmark",
                        "maxLength": 255
                    },
                    "slug": {
                        "description": "Slug of the book.",
                        "type": "string",
                        "pattern": "^[a-z0-9-]+$",
                        "example": "the-tragical-history-of-hamlet-prince-of-denmark",
                        "maxLength": 255
                    },
                    "description": {
                        "description": "Description of the book.",
                        "type": "string",
                        "example": "The story follows Prince Hamlet of Denmark, who returns home to find his father (the King) dead and his uncle, Claudius, newly crowned and married to Hamlet's mother, Gertrude. After the Ghost of his father appears and reveals that Claudius murdered him, Hamlet is consumed by a quest for revenge.",
                        "nullable": true,
                        "maxLength": 5000
                    },
                    "publication_date": {
                        "description": "Publication date of the book.",
                        "type": "string",
                        "format": "date",
                        "example": "1603-01-01",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "BookResponse": {
                "title": "Book Response",
                "description": "Data of a specific book.",
                "required": [
                    "id",
                    "title",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the book.",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the book.",
                        "type": "string",
                        "example": "The Tragical History of Hamlet, Prince of Denmark"
                    },
                    "slug": {
                        "description": "Slug of the book.",
                        "type": "string",
                        "example": "the-tragical-history-of-hamlet-prince-of-denmark"
                    },
                    "description": {
                        "description": "Description of the book.",
                        "type": "string",
                        "example": "The story follows Prince Hamlet of Denmark, who returns home to find his father (the King) dead and his uncle, Claudius, newly crowned and married to Hamlet's mother, Gertrude. After the Ghost of his father appears and reveals that Claudius murdered him, Hamlet is consumed by a quest for revenge.",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Image url of the book.",
                        "type": "string",
                        "format": "url",
                        "example": "http://loclhost:8080/storage/books/book1.png",
                        "nullable": true
                    },
                    "publication_date": {
                        "description": "Publication date of the book.",
                        "type": "string",
                        "format": "date",
                        "example": "1603-01-01",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "BookFullResponse": {
                "title": "Book Full Response",
                "description": "Data of a specific book.",
                "required": [
                    "id",
                    "title",
                    "slug",
                    "authors"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the book.",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the book.",
                        "type": "string",
                        "example": "The Tragical History of Hamlet, Prince of Denmark"
                    },
                    "slug": {
                        "description": "Slug of the book.",
                        "type": "string",
                        "example": "the-tragical-history-of-hamlet-prince-of-denmark"
                    },
                    "description": {
                        "description": "Description of the book.",
                        "type": "string",
                        "example": "The story follows Prince Hamlet of Denmark, who returns home to find his father (the King) dead and his uncle, Claudius, newly crowned and married to Hamlet's mother, Gertrude. After the Ghost of his father appears and reveals that Claudius murdered him, Hamlet is consumed by a quest for revenge.",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Image url of the book.",
                        "type": "string",
                        "format": "url",
                        "example": "http://loclhost:8080/storage/books/book1.png",
                        "nullable": true
                    },
                    "publication_date": {
                        "description": "Publication date of the book.",
                        "type": "string",
                        "format": "date",
                        "example": "1603-01-01",
                        "nullable": true
                    },
                    "authors": {
                        "description": "Authors of the book.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuthorResponse"
                        },
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}